requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
torch>=2.0.0
mlflow>=2.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
pyyaml>=6.0
scipy>=1.11.0
pyarrow>=13.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
lightgbm>=4.0.0
xgboost>=2.0.0
optuna>=3.4.0
streamlit>=1.32.0
altair>=5.0.0
requests>=2.31.0
